refactor(schema): decompose index
