+++ packages/schema/src/main/ts/utils.ts
@@ -1,0 +1,3 @@
+export const isObject = (value: any): boolean => value !== null && typeof value === 'object'
+
+export const joinPath = (a: string, b: string): string => a ? `${a}.${b}` : b
